refactor: resolve react lint warnings
