+++ src/components/Editor/EditorHeader/BackButton.tsx
@@ -3,7 +3,7 @@
  */
 
 import { useNavigate } from 'react-router-dom';
-import { useEditorContext } from '@/contexts/EditorContext';
+import { useEditorContext } from '@/contexts/editor-context';
 import { Button } from '@/components/ui/button';
 import { ArrowLeft } from 'lucide-react';
